+++ Sources/ProjectDescription/TargetDependency.swift
@@ -2,11 +2,62 @@ import Foundation
 
 // MARK: - TargetDependency
 
-public enum TargetDependency: Codable {
+/// Dependency status used by `.sdk` target dependencies
+public enum SDKStatus: String {
+    /// Required dependency
+    case required
+
+    /// Optional dependency (weakly linked)
+    case optional
+}
+
+/// Defines the target dependencies supported by Tuist
+public enum TargetDependency: Codable, Equatable {
+    /// Dependency on another target within the same project
+    ///
+    /// - Parameters:
+    ///   - name: Name of the target to depend on
     case target(name: String)
+
+    /// Dependency on a target within another project
+    ///
+    /// - Parameters:
+    ///   - target: Name of the target to depend on
+    ///   - path: Relative path to the other project directory
     case project(target: String, path: String)
+
+    /// Dependency on a prebuilt framework
+    ///
+    /// - Parameters:
+    ///   - path: Relative path to the prebuilt framework
     case framework(path: String)
+
+    /// Dependency on prebuilt library
+    ///
+    /// - Parameters:
+    ///   - path: Relative path to the prebuilt library
+    ///   - publicHeaders: Relative path to the library's public headers directory
+    ///   - swiftModuleMap: Relative path to the library's swift module map file
     case library(path: String, publicHeaders: String, swiftModuleMap: String?)
+
+    /// Dependency on system library or framework
+    ///
+    /// - Parameters:
+    ///   - name: Name of the system library or framework (including extension)
+    ///            e.g. `ARKit.framework`, `libc++.tbd`
+    ///   - status: The dependency status (optional dependencies are weakly linked)
+    case sdk(name: String, status: SDKStatus)
+
+    /// Dependency on system library or framework
+    ///
+    /// - Parameters:
+    ///   - name: Name of the system library or framework (including extension)
+    ///            e.g. `ARKit.framework`, `libc++.tbd`
+    ///
+    /// Note: Defaults to using a `required` dependency status
+    public static func sdk(name: String) -> TargetDependency {
+        return .sdk(name: name, status: .required)
+    }
 
     public var typeName: String {
         switch self {
@@ -18,9 +69,15 @@ public enum TargetDependency: Codable {
             return "framework"
         case .library:
             return "library"
+        case .sdk:
+            return "sdk"
         }
     }
 }
+
+// MARK: - SDKStatus (Coding)
+
+extension SDKStatus: Codable {}
 
 // MARK: - TargetDependency (Coding)
 
@@ -36,6 +93,7 @@ extension TargetDependency {
         case path
         case publicHeaders = "public_headers"
         case swiftModuleMap = "swift_module_map"
+        case status
     }
 
     public init(from decoder: Decoder) throws {
@@ -63,6 +121,10 @@ extension TargetDependency {
                 swiftModuleMap: try container.decodeIfPresent(String.self, forKey: .swiftModuleMap)
             )
 
+        case "sdk":
+            self = .sdk(name: try container.decode(String.self, forKey: .name),
+                        status: try container.decode(SDKStatus.self, forKey: .status))
+
         default:
             throw CodingError.unknownType(type)
         }
@@ -85,6 +147,9 @@ extension TargetDependency {
             try container.encode(path, forKey: .path)
             try container.encode(publicHeaders, forKey: .publicHeaders)
             try container.encodeIfPresent(swiftModuleMap, forKey: .swiftModuleMap)
+        case let .sdk(name, status):
+            try container.encode(name, forKey: .name)
+            try container.encode(status, forKey: .status)
         }
     }
 }