+++ Tests/TuistGeneratorTests/Generator/ProjectFileElementsTests.swift
@@ -69,7 +69,7 @@ final class ProjectFileElementsTests: XCTestCase {
 
         // Then
         let projectGroup = groups.main.group(named: "Project")
-        XCTAssertEqual(projectGroup?.debugChildPaths, [
+        XCTAssertEqual(projectGroup?.flattenedChildren, [
             "myfolder/resources/a.png",
         ])
     }
@@ -87,7 +87,7 @@ final class ProjectFileElementsTests: XCTestCase {
 
         // Then
         let projectGroup = groups.main.group(named: "Project")
-        XCTAssertEqual(projectGroup?.debugChildPaths, [
+        XCTAssertEqual(projectGroup?.flattenedChildren, [
             "my.folder/resources/a.png",
         ])
     }
@@ -106,7 +106,7 @@ final class ProjectFileElementsTests: XCTestCase {
 
         // Then
         let projectGroup = groups.main.group(named: "Project")
-        XCTAssertEqual(projectGroup?.debugChildPaths, [
+        XCTAssertEqual(projectGroup?.flattenedChildren, [
             "myfolder/resources/generated_images",
         ])
     }
@@ -124,7 +124,7 @@ final class ProjectFileElementsTests: XCTestCase {
 
         // Then
         let projectGroup = groups.main.group(named: "Project")
-        XCTAssertEqual(projectGroup?.debugChildPaths, [
+        XCTAssertEqual(projectGroup?.flattenedChildren, [
             "another/path/resources/a.png",
         ])
     }
@@ -142,7 +142,7 @@ final class ProjectFileElementsTests: XCTestCase {
 
         // Then
         let projectGroup = groups.main.group(named: "Project")
-        XCTAssertEqual(projectGroup?.debugChildPaths, [
+        XCTAssertEqual(projectGroup?.flattenedChildren, [
             "myfolder/resources/assets.xcassets",
         ])
     }
@@ -170,7 +170,7 @@ final class ProjectFileElementsTests: XCTestCase {
 
         // Then
         let projectGroup = groups.main.group(named: "Project")
-        XCTAssertEqual(projectGroup?.debugChildPaths, [
+        XCTAssertEqual(projectGroup?.flattenedChildren, [
             "myfolder/resources/assets.xcassets",
         ])
     }
@@ -200,7 +200,7 @@ final class ProjectFileElementsTests: XCTestCase {
 
         // Then
         let projectGroup = groups.main.group(named: "Project")
-        XCTAssertEqual(projectGroup?.debugChildPaths, [
+        XCTAssertEqual(projectGroup?.flattenedChildren, [
             "resources/App.strings/en",
             "resources/App.strings/fr",
             "resources/Extension.strings/en",
@@ -585,6 +585,36 @@ final class ProjectFileElementsTests: XCTestCase {
                                                      sourceRootPath: AbsolutePath("/a/b/c/project"))
         XCTAssertEqual(got, RelativePath("../../../framework"))
     }
+
+    func test_generateDependencies_sdks() throws {
+        // Given
+        let pbxproj = PBXProj()
+        let project = Project.test()
+        let sourceRootPath = AbsolutePath("/a/project/")
+        let groups = ProjectGroups.generate(project: project,
+                                            pbxproj: pbxproj,
+                                            sourceRootPath: sourceRootPath)
+
+        let sdk = try SDKNode(name: "ARKit.framework", status: .required)
+
+        // When
+        try subject.generate(dependencies: [sdk],
+                             path: sourceRootPath,
+                             groups: groups, pbxproj: pbxproj,
+                             sourceRootPath: sourceRootPath,
+                             filesGroup: .group(name: "Project"))
+
+        // Then
+        XCTAssertEqual(groups.frameworks.flattenedChildren, [
+            "ARKit.framework",
+        ])
+
+        let sdkElement = subject.sdks[sdk.path]
+        XCTAssertNotNil(sdkElement)
+        XCTAssertEqual(sdkElement?.sourceTree, .sdkRoot)
+        XCTAssertEqual(sdkElement?.path, sdk.path.relative(to: "/").pathString)
+        XCTAssertEqual(sdkElement?.name, sdk.path.basename)
+    }
 }
 
 private extension PBXGroup {
@@ -597,11 +627,11 @@ private extension PBXGroup {
     ///    -- D
     /// Would return:
     ///         ["A/B", "A/C/D"]
-    var debugChildPaths: [String] {
+    var flattenedChildren: [String] {
         return children.flatMap { (element: PBXFileElement) -> [String] in
             switch element {
             case let group as PBXGroup:
-                return group.debugChildPaths.map { group.nameOrPath + "/" + $0 }
+                return group.flattenedChildren.map { group.nameOrPath + "/" + $0 }
             default:
                 return [element.nameOrPath]
             }